Added Consumer for AMQP AVRO messages from RabbitMQ
Added RabbitMQ AMQP docker container as a broker.
Sent the AMQP messages using the `camel-rabbitmq-kafka-sink-connector`
to RabbitMQ.

Wrote a consumer which decodes the AVRO messages.
This AVRO consumer use the Confluent Schema Registry.
Added some test table message decoding.

ToDo: The AMQP message needs to be added with headers mentioning the
- Subject
- Version ID
With these it is possible to decode the AVRO binary message in the payload.